+++ .idea/PythonExtensions.iml
@@ -9,8 +9,9 @@
       <excludeFolder url="file://$MODULE_DIR$/PythonExtensions-1.0.8" />
       <excludeFolder url="file://$MODULE_DIR$/PythonExtensions.egg-info" />
       <excludeFolder url="file://$MODULE_DIR$/PythonExtensions-1.0.11" />
+      <excludeFolder url="file://$MODULE_DIR$/venv" />
     </content>
-    <orderEntry type="jdk" jdkName="Python 3.9" jdkType="Python SDK" />
+    <orderEntry type="jdk" jdkName="Python 3.10 (PythonExtensions)" jdkType="Python SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
   </component>
 </module>
