+++ PythonExtensions/tk/Enumerations.py
